Split ClassFile.hs into smaller modules
- also added breaking tests
